Refactor: home page 구현 부분을 pageComponents/home에서 components/home으로 이동
